Moved the data_coordinator class under the data_ingestion hierarchy.
