+++ include/lbann/CMakeLists.txt
@@ -34,7 +34,6 @@ set_full_path(THIS_DIR_HEADERS
 
 # Add the subdirectories
 add_subdirectory(callbacks)
-add_subdirectory(data_coordinator)
 add_subdirectory(data_ingestion)
 add_subdirectory(io)
 add_subdirectory(layers)
