+++ src/CMakeLists.txt
@@ -32,7 +32,6 @@ set_full_path(THIS_DIR_SOURCES
 
 # Add the subdirectories
 add_subdirectory(callbacks)
-add_subdirectory(data_coordinator)
 add_subdirectory(data_ingestion)
 add_subdirectory(io)
 add_subdirectory(layers)
