+++ src/callbacks/check_dataset.cpp
@@ -26,7 +26,7 @@
 
 #include "lbann/callbacks/check_dataset.hpp"
 #include "lbann/comm_impl.hpp"
-#include "lbann/data_coordinator/data_coordinator.hpp"
+#include "lbann/data_ingestion/coordinator/data_coordinator.hpp"
 #include "lbann/execution_algorithms/execution_context.hpp"
 #include "lbann/execution_algorithms/sgd_execution_context.hpp"
 #include "lbann/layers/io/input_layer.hpp"
