+++ src/callbacks/debug_io.cpp
@@ -29,7 +29,7 @@
 #include "lbann/callbacks/debug_io.hpp"
 
 #include "lbann/base.hpp"
-#include "lbann/data_coordinator/data_coordinator.hpp"
+#include "lbann/data_ingestion/coordinator/data_coordinator.hpp"
 #include "lbann/execution_algorithms/sgd_execution_context.hpp"
 #include "lbann/models/model.hpp"
 #include "lbann/trainers/trainer.hpp"
