+++ src/callbacks/variable_minibatch.cpp
@@ -28,7 +28,7 @@
 
 #include "lbann/callbacks/variable_minibatch.hpp"
 
-#include "lbann/data_coordinator/data_coordinator.hpp"
+#include "lbann/data_ingestion/coordinator/data_coordinator.hpp"
 #include "lbann/execution_algorithms/sgd_execution_context.hpp"
 #include "lbann/layers/io/input_layer.hpp"
 #include "lbann/models/model.hpp"
