+++ src/data_ingestion/coordinator/buffered_data_coordinator.cpp
@@ -25,7 +25,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include "lbann/comm_impl.hpp"
-#include "lbann/data_coordinator/buffered_data_coordinator_impl.hpp"
+#include "lbann/data_ingestion/coordinator/buffered_data_coordinator_impl.hpp"
 #include "lbann/data_ingestion/data_reader.hpp"
 #include "lbann/data_ingestion/data_store_conduit.hpp"
 #include "lbann/data_ingestion/infrastructure/data_packer.hpp"
