+++ src/data_ingestion/coordinator/data_coordinator.cpp
@@ -25,7 +25,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include <lbann/comm_impl.hpp>
-#include <lbann/data_coordinator/data_coordinator.hpp>
+#include <lbann/data_ingestion/coordinator/data_coordinator.hpp>
 #include <lbann/data_ingestion/data_reader.hpp>
 #include <lbann/data_ingestion/data_store_conduit.hpp>
 #include <lbann/execution_algorithms/execution_context.hpp>
