+++ src/models/model.cpp
@@ -31,7 +31,7 @@
 #include "lbann/callbacks/checkpoint.hpp"
 #include "lbann/callbacks/save_model.hpp"
 #include "lbann/comm_impl.hpp"
-#include "lbann/data_coordinator/data_coordinator.hpp"
+#include "lbann/data_ingestion/coordinator/data_coordinator.hpp"
 #include "lbann/data_ingestion/data_store_conduit.hpp"
 #include "lbann/execution_algorithms/execution_context.hpp"
 #include "lbann/io/persist.hpp"
